+++ BPt/extensions/__init__.py
@@ -1,5 +1,9 @@
-from .FeatSelectors import FeatureSelector
+from .feat_selectors import FeatureSelector
 from .loaders import Identity, get_loader_pipe
+
+from .mlp import MLPRegressor_Wrapper, MLPClassifier_Wrapper
+from .scalers import Winsorizer
+from .residualizer import LinearResidualizer
 
 try:
     from .loaders import SingleConnectivityMeasure
@@ -12,11 +16,6 @@ try:
 except ImportError:
     class ThresholdNetworkMeasures():
         pass
-
-from .MLP import MLPRegressor_Wrapper, MLPClassifier_Wrapper
-from .Scalers import Winsorizer
-from .residualizer import LinearResidualizer
-
 
 __all__ = ['FeatureSelector', 'get_loader_pipe',
            'Identity', 'SingleConnectivityMeasure',
