+++ BPt/extensions/tests/test_feat_selectors.py
@@ -1,5 +1,5 @@
 import numpy as np
-from ..FeatSelectors import FeatureSelector
+from ..feat_selectors import FeatureSelector
 
 
 def basic_test():
